pyproject.toml
[project]
name = "comfyui-ollamagemini"
description = "ComfyUI extension for Ollama, Gemini, OpenAI, Claude, and Qwen with video and audio support"
version = "2.7"
license = { file = "LICENSE" }
dependencies = [
    "google-generativeai",
    "google-genai>=1.50.0",
    "requests>=2.31.0",
    "openai>=1.3.0",
    "anthropic>=0.8.0",
    "dashscope>=1.13.6",
    "vtracer>=0.6.0",
    "numpy>=1.24.0",
    "Pillow>=10.0.0",
    "scipy>=1.10.0",
    "opencv-python>=4.6.0",
    "transformers==4.55.4",
    "torch",
    "torchaudio",
]

[project.urls]
Repository = "https://github.com/al-swaiti/ComfyUI-OllamaGemini"
#  Used by Comfy Registry https://comfyregistry.org

[tool.comfy]
PublisherId = "abdallaalsw80"
DisplayName = "ComfyUI-OllamaGemini"
Icon = "https://raw.githubusercontent.com/al-swaiti/ComfyUI-OllamaGemini/refs/heads/main/ollamagemini.png"
